gemini_trade_service/requirements.txt
flask==3.0.0
google-genai==0.2.1
requests==2.31.0
pandas==2.1.4

dhanhq==2.0.2
python-dotenv==1.0.0
yfinance==0.2.32
beautifulsoup4==4.12.2
aiohttp==3.9.1
numpy==1.24.4
